+++ client/index.js
@@ -7,7 +7,7 @@ import { render } from '@wordpress/element';
  * Internal dependencies
  */
 import './stylesheets/index.scss';
-import Navigation from './_navigation';
+import Navigation from './navigation';
 
 const navigationRoot = document.getElementById(
 	'woocommerce-embedded-navigation'
